# Use native isoformat/fromisoformat