--- a/marshmallow_recipe/fields.py
+++ b/marshmallow_recipe/fields.py
@@ -645,6 +645,16 @@ if _MARSHMALLOW_VERSION_MAJOR >= 3:
                 else {}
             ),
         }
+
+        @staticmethod
+        def __fromisoformat_py310_fix(
+            format: str, value: str
+        ) -> datetime.datetime:  # pyright: ignore[reportUnusedFunction]
+            try:
+                return m.fields.DateTime.DESERIALIZATION_FUNCS[format](value)  # type: ignore
+            except ValueError:
+                return datetime.datetime.strptime(value, "%Y-%m-%d")
+
         DESERIALIZATION_FUNCS = {
             **m.fields.DateTime.DESERIALIZATION_FUNCS,  # type: ignore
             **(
@@ -653,7 +663,10 @@ if _MARSHMALLOW_VERSION_MAJOR >= 3:
                     "iso8601": datetime.datetime.fromisoformat,
                 }
                 if sys.version_info >= (3, 11)
-                else {}
+                else {
+                    "iso": lambda x: DateTimeFieldV3.__fromisoformat_py310_fix("iso", x),
+                    "iso8601": lambda x: DateTimeFieldV3.__fromisoformat_py310_fix("iso8601", x),
+                }
             ),
         }
 
@@ -877,6 +890,16 @@ else:
                 else {}
             ),
         }
+
+        @staticmethod
+        def __fromisoformat_py310_fix(
+            format: str, value: str
+        ) -> datetime.datetime:  # pyright: ignore[reportUnusedFunction]
+            try:
+                return m.fields.DateTime.DATEFORMAT_DESERIALIZATION_FUNCS[format](value)  # type: ignore
+            except ValueError:
+                return datetime.datetime.strptime(value, "%Y-%m-%d")
+
         DATEFORMAT_DESERIALIZATION_FUNCS = {
             **m.fields.DateTime.DATEFORMAT_DESERIALIZATION_FUNCS,  # type: ignore
             **(
@@ -885,7 +908,10 @@ else:
                     "iso8601": datetime.datetime.fromisoformat,
                 }
                 if sys.version_info >= (3, 11)
-                else {}
+                else {
+                    "iso": lambda x: DateTimeFieldV2.__fromisoformat_py310_fix("iso", x),
+                    "iso8601": lambda x: DateTimeFieldV2.__fromisoformat_py310_fix("iso8601", x),
+                }
             ),
         }